+++ packages/main/src/components/ProgressIndicator/index.tsx
@@ -1,9 +1,8 @@
 import { StyleClassHelper } from '@ui5/webcomponents-react-base/lib/StyleClassHelper';
 import { usePassThroughHtmlProps } from '@ui5/webcomponents-react-base/lib/usePassThroughHtmlProps';
-import { ContentDensity } from '@ui5/webcomponents-react/lib/ContentDensity';
 import { ValueState } from '@ui5/webcomponents-react/lib/ValueState';
 import React, { FC, forwardRef, Ref, useMemo } from 'react';
-import { createUseStyles, useTheme } from 'react-jss';
+import { createUseStyles } from 'react-jss';
 import { CommonProps } from '../../interfaces/CommonProps';
 import { JSSTheme } from '../../interfaces/JSSTheme';
 import styles from './ProgressIndicator.jss';
@@ -34,7 +33,7 @@ export interface ProgressIndicatorPropTypes extends CommonProps {
   state?: ValueState;
 }
 
-const useStyles = createUseStyles<keyof ReturnType<typeof styles>>(styles, { name: 'ProgressIndicator' });
+const useStyles = createUseStyles<JSSTheme, keyof ReturnType<typeof styles>>(styles, { name: 'ProgressIndicator' });
 
 /**
  * <code>import { ProgressIndicator } from '@ui5/webcomponents-react/lib/ProgressIndicator';</code>
@@ -51,12 +50,6 @@ const ProgressIndicator: FC<ProgressIndicatorPropTypes> = forwardRef(
     const progressBarTextClasses = StyleClassHelper.of(classes.progressBarText);
 
     const progressBarStyle = { flexBasis: `${percentValue}%` };
-
-    // change content density
-    const theme = useTheme() as JSSTheme;
-    if (theme.contentDensity === ContentDensity.Compact) {
-      wrapperClasses.put(classes.compact);
-    }
 
     // change text color based on percent value
     if (percentValue <= 50) {